Updated with changes from OmniGroup svn [r246662:250963]
General:
- Enabled ObjC modules.
- Fixed remaining shadow warnings.
- Enabled support for Swift embedding.
- Switched from OMNI_BUNDLE_IDENTIFIER to PRODUCT_BUNDLE_IDENTIFIER, now that the latter exists.
- Cleaned up version numbers in plists that the App Store(s) submission process can complain about.
- Converted more code to ARC (and switched some delegates to weak).
- Added generics annotations in some places.
- Added nullability annotations in some places.
- Updated localizations.

Configurations:
- Enable ObjC modules for all frameworks, by default.
- Removed OMNI_BUNDLE_IDENTIFIER as a -D flag, since that causes excessive precompilation of module headers.
- Unifiy OMNI_BUILDING_BUNDLE and OMNI_BUILDING_FRAMEWORK into OMNI_BUILDING_FRAMEWORK_OR_BUNDLE for the same reason.
- Enable EMBEDDED_CONTENT_CONTAINS_SWIFT for applications, by default.
- Enable GCC_WARN_SHADOW globally.
- Removed -Wno-error-nonnull and fixed remaining warnings of that kind (or added bugs to look at them).
- Added Touch-XCTest-* configurations for unit test bundles.

OmniBase:
- Since adding -DOMNI_BUNDLE_IDENTIFIER="whatever" doesn't play well with modules, OMNI_BUNDLE is now a macro that looks up the bundle via the dlopen APIs.
- Added OBLogS for Swift callers.
- Speed up the runtime checks OmniBase does (and avoid redundant checks on first launch).
- Avoid checking for API mismatches between system classes by default (not our problem, but can still be useful when tracking down problems...).

OmniFoundation:
- Renamed OFController -addObserver:, and friends, to use 'StatusObserver' to avoid conflicting with private methods added to Foundation in 10.11.
- Added OFCreatePropertyListFromNSData()
- Split OFStringHash_djb2 into some alternate input formats, and use CFStringInitInlineBuffer().
- Added OFSelectionSet, which is most of what was OIInspectionSet.
- Added +[OFVersionNumber mainBundleVersionNumber].
- Added a Swift wrapper of OFUTI.
- Added more supportting functions and tests for encryption.
- Added OFObservation, a wrapper for KVO operations (not fully fleshed out or used anywhere for real yet).
- Added OFRangeContainsRange().
- Updated OFDedicatedThreadScheduler to observe NSSystemClockDidChangeNotification instead of having subclasses call its -scheduleEvents.
- Fix encoding of surrogate pair character entities.

OmniAppKit:
- Added support for switching theme plists.
- Promoted +handleChangePreferenceURL: out of OmniFocus into OAController.
- Did a better job of documenting the OAToolbarHelper.

OmniDAV:
- Allow clients to customize the first component of the User-Agent header added the marketing version after the build version.
- More WIP for encryption.
- Instead of rejecting a call to -changedValues if we're looking up a value on the other end of a to-one relationship that's been inserted from an undo, fall back on -primitiveValueForProperty: to look up a real ODOObject value (instead of the primary key string that we have from the undo snapshot) and return it.

OmniDocumentStore:
- Avoid a potential crash when moving a large number of items between scopes.

OmniFileStore:
- More progress on encryption.

OmniInspector:
- Fix logic for auto-selecting appropriate inspector tabs.
- Improve handling of scrollbars in inspectors.
- Make OIInspectionSet subclass OFSelectionSet, which most of the code moving to OmniFoundation.

OmniSoftwareUpdate:
- Converted OSUAvailableUpdateController to Base localization and autolayout.
- Don't show the list of releases and release notes unless there is an available update.

OmniUI/iPad:
- Updated inspectors to use a UIStackView and constraints instead of custom code.
- Update help viewer to use OUIWebViewController.
- Add image support to OUITabBar and OUITabBarButton.
- Add support for customizing help location for apps.
- Truncate titles in OUITabBarButton when needed.
- Make fewer assumptions about the keyboard docking behavior in OUIKeyboardNotifier.
- Add OUIUploadController for managing the workflow of uploading documents.

OmniUnzip:
- Move the definitions of NOCRYPT and VERSIONMADEBY to the prefix header instead of defining them on the command line while building. This avoids the creation of extra module precompiled headers.

+++ Frameworks/OmniAppKit/OpenStepExtensions.subproj/OAFontDescriptor.m
@@ -159,9 +159,9 @@ static NSInteger _weightToFontManagerWeight(CGFloat weight)
     if (weight < -0.8)
         return 1;
     else if (weight <= 0.2)
-        return round(5.0 * weight + 5.0 );
+        return (NSInteger)round(5.0 * weight + 5.0 );
     else if (weight <= 0.8)
-        return round(8.0 * weight / 0.6 + 10.0 / 3.0);
+        return (NSInteger)round(8.0 * weight / 0.6 + 10.0 / 3.0);
     else
         return 14;
 }
